Remove now migrated datetime tests
Avoid using datetime-related Cypher functions in unrelated
integration tests since the driver is now not usable until the
5.0 server supports UTC DateTime values
